Create NestedClass-MetroStationClass
In this Code, I have made a class within a class. This is known as nested Class. I have made a Class for Metro Station and Ticket Purchasing system.
I have already defined the price of ticket and take an amount from the user and tell him either the ticket is purchased or not and also his remaining balance.